pyproject.toml
[project]
name = "llm-proxy-cli"
version = "0.2.0"
description = "Communication with LLM with CLI and Security Features"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "llm-guard>=0.3.16",
    "PyYAML>=6.0",
]

[dependency-groups]
dev = [
    "huggingface-hub[hf-xet]>=0.34.3",
    "pip>=25.2",
    "ruff>=0.12.7",
]
